+++ .gitignore
@@ -8,8 +8,8 @@ node_modules
 # testing
 /coverage
 
-# production
-/build
+# build target
+/dist
 
 # misc
 .DS_Store
