+++ package.json
@@ -61,12 +61,14 @@
     "webpack": "^5.90.3",
     "webpack-cli": "^5.1.4",
     "webpack-dev-middleware": "^7.0.0",
-    "webpack-dev-server": "^4.15.1"
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^6.0.1"
   },
   "scripts": {
-    "start": "node server",
-    "devserver": "webpack-dev-server",
-    "build": "webpack",
+    "start": "npm run build && npm run serve",
+    "serve": "node server",
+    "devserver": "webpack-dev-server --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js",
     "prettier": "prettier --write ."
   },
   "eslintConfig": {
