+++ webpack.common.js
@@ -1,0 +1,111 @@
+const path = require('path');
+const webpack = require('webpack');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+require('dotenv').config();
+
+module.exports = {
+  entry: {
+    main: [path.join(process.cwd(), 'src/index.js')],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: ['@babel/preset-env', '@babel/preset-react'],
+          },
+        },
+      },
+      {
+        test: /\.css$/,
+        use: ['style-loader', 'css-loader'],
+      },
+      {
+        test: /\.svg$/,
+        type: 'asset', // Use asset module type for SVG
+        parser: {
+          dataUrlCondition: {
+            maxSize: 8 * 1024, // 8kb
+          },
+        },
+      },
+      {
+        test: /\.(jpg|png|gif)$/,
+        type: 'asset/resource', // Use asset module type for images
+      },
+      {
+        test: /\.html$/,
+        use: {
+          loader: 'html-loader',
+          options: {
+            sources: {
+              urlFilter: (attribute, value) => !(/environment.js$/.test(value))
+            }
+          }
+        }
+      },
+      {
+        test: /\.m?js/,
+        resolve: {
+          fullySpecified: false,
+        },
+      },
+    ],
+  },
+  output: {
+    path: path.resolve(process.cwd(), 'dist'),
+    publicPath: '/',
+    filename: '[name].[contenthash].js',
+    chunkFilename: '[name].[contenthash].chunk.js',
+    clean: true, // Automatically clean the output directory before each build
+  },
+  optimization: {
+    moduleIds: 'deterministic', // Enable consistent hashing for long term caching
+    runtimeChunk: 'single', // Create a runtime file to be shared for all generated chunks
+    splitChunks: {
+      chunks: 'all',
+      maxInitialRequests: Infinity,
+      minSize: 0,
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name (module) {
+            const packageName = module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1];
+            return `npm.${packageName.replace('@', '')}`;
+          },
+        },
+      },
+    },
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      inject: true,
+      publicPath: '.',
+      template: 'src/index.html',
+      favicon: 'src/favicon.ico',
+      manifest: 'src/manifest.json',
+    }),
+    new webpack.ProvidePlugin({
+      process: 'process/browser',
+      Buffer: ['buffer', 'Buffer'],
+    }),
+  ],
+  resolve: {
+    modules: ['node_modules', 'src'],
+    extensions: ['.js', '.jsx', '.react.js'],
+    mainFields: ['browser', 'module', 'main'],
+    fallback: {
+      crypto: require.resolve('crypto-browserify'),
+      stream: require.resolve('stream-browserify'),
+      vm: false,
+    },
+  },
+  target: ['web', 'es5'],
+  performance: {
+    hints: false,
+  },
+};
